requirements.txt
prefect
python-dotenv
undetected-chromedriver
markdownify
langchain
networkx
pandas
tqdm
